# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Finance Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: Get transactions with filters
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    put:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionInput"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [cash, card, wallet, savings, investment]
        currency:
          type: string
        initialBalance:
          type: number
        currentBalance:
          type: number
        color:
          type: string
        icon:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, currency, initialBalance, currentBalance, createdAt]
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        accountId:
          type: integer
        categoryId:
          type: integer
        subcategory:
          type: string
        date:
          type: string
          format: date
        comment:
          type: string
        receiptUrl:
          type: string
        isTransfer:
          type: boolean
        createdAt:
          type: string
          format: date-time
        account:
          $ref: "#/components/schemas/Account"
        category:
          $ref: "#/components/schemas/Category"
      required: [id, type, amount, accountId, categoryId, date, createdAt]
    CreateTransactionInput:
      type: object
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        accountId:
          type: integer
        categoryId:
          type: integer
        subcategory:
          type: string
        date:
          type: string
          format: date
        comment:
          type: string
        receiptUrl:
          type: string
      required: [type, amount, accountId, categoryId, date]
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        totalIncome:
          type: number
        totalExpense:
          type: number
      required: [transactions, total, totalIncome, totalExpense]
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [income, expense, both]
        color:
          type: string
        icon:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, color, icon, isDefault, createdAt]
...